requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==4.0.3
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
docstring_parser==0.16
faiss-cpu==1.10.0
filelock==3.17.0
flask==3.1.0
frozenlist==1.5.0
fsspec==2025.2.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
idna==3.10
itsdangerous==2.2.0
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.34
langchain-google-vertexai==2.0.13
langchain-huggingface==0.1.2
langchain-mistralai==0.2.6
langchain-openai==0.3.4
langchain-text-splitters==0.3.6
langgraph==0.2.70
langgraph-checkpoint==2.0.12
langgraph-sdk==0.1.51
langsmith==0.3.8
marshmallow==3.26.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
openai==1.61.1
orjson==3.10.15
pillow==11.1.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
regex==2024.11.6
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==3.4.1
shapely==2.0.7
SQLAlchemy==2.0.38
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
typing-inspect==0.9.0
webencodings==0.5.1
werkzeug==3.1.3
yarl==1.18.3
zstandard==0.23.0
